+++ lib/elixir/test/elixir/kernel/typespec_test.exs
@@ -3,8 +3,6 @@ Code.require_file "../test_helper.exs", __DIR__
 defmodule Kernel.TypespecTest do
   use ExUnit.Case, async: true
 
-  defrecord Rec, [:first, :last]
-
   # This macro allows us to focus on the result of the
   # definition and not on the hassles of handling test
   # module
@@ -208,14 +206,6 @@ defmodule Kernel.TypespecTest do
     assert {:mytype, {:type, _, :union, [{:type, _, :integer, []},
              {:remote_type, _, [{:atom, _, :elixir}, {:atom, _, :char_list}, []]},
              {:type, _, :atom, []}]}, []} = spec
-  end
-
-  test "@type with an access macro" do
-    spec = test_module do
-      @type mytype :: Rec[first: integer]
-    end
-    assert {:mytype, {:type, _, :tuple,
-             [{:atom, _, Rec}, {:type, _, :integer, []}, {:type, _, :any, []}]}, []} = spec
   end
 
   test "@type with keywords" do
@@ -436,33 +426,6 @@ defmodule Kernel.TypespecTest do
       ast = Kernel.Typespec.type_to_ast(type)
       assert Macro.to_string(quote do: @type unquote(ast)) == Macro.to_string(definition)
     end
-  end
-
-  test "type_to_ast for records" do
-    record_type = {{:record, :my_record},
-                    [
-                      {:typed_record_field,
-                        {:record_field, 0, {:atom, 0, :field1}},
-                        {:type, 0, :atom, []}},
-                      {:typed_record_field,
-                        {:record_field, 0, {:atom, 0, :field2}},
-                        {:type, 0, :integer, []}},
-                    ],
-                    []}
-    assert Kernel.Typespec.type_to_ast(record_type) ==
-      {:::, [], [
-        {:my_record, [], []},
-        {:{}, [], [:my_record,
-          {:::, [line: 0], [
-            {:field1, 0, nil},
-            {:atom, [line: 0], []}
-          ]},
-          {:::, [line: 0], [
-            {:field2, 0, nil},
-            {:integer, [line: 0], []}
-          ]}
-        ]}
-      ]}
   end
 
   test "type_to_ast for paren_type" do